move _clean_notebooks outside of weldx package
since this tool is only a developer utility.
